Backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers
Pillow==12.0.0
gunicorn
whitenoise
